move init_django into utils
