+++ scripts/utils.py
@@ -5,6 +5,8 @@ import uuid
 import datetime
 import yaml
 import yamlordereddictloader
+import django
+from django import conf
 from collections import defaultdict
 from yaml.representer import Representer
 import openstates_metadata as metadata
@@ -101,3 +103,27 @@ def legacy_districts(**kwargs):
     for d in metadata.lookup(**kwargs).legacy_districts:
         legacy_districts[d.chamber_type].append(d.name)
     return legacy_districts
+
+
+def init_django():  # pragma: no cover
+    conf.settings.configure(
+        conf.global_settings,
+        SECRET_KEY="not-important",
+        DEBUG=False,
+        INSTALLED_APPS=(
+            "django.contrib.contenttypes",
+            "opencivicdata.core.apps.BaseConfig",
+            "opencivicdata.legislative.apps.BaseConfig",
+        ),
+        DATABASES={
+            "default": {
+                "ENGINE": "django.contrib.gis.db.backends.postgis",
+                "NAME": os.environ["PGDATABASE"],
+                "USER": os.environ["PGUSER"],
+                "PASSWORD": os.environ["PGPASSWORD"],
+                "HOST": os.environ["PGHOST"],
+            }
+        },
+        MIDDLEWARE_CLASSES=(),
+    )
+    django.setup()
